+++ extensions/hammerspoon/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Hammerspoon Changelog
 
+## [Improvements] 1.2.0 - {PR_MERGE_DATE}
+
+- Browse Documentation command
+- add option to not open console on restart
+
 ## [Improvements] 1.1.0 - 2025-06-04
 
 - add Open Documentation command